fix: disable creation of WriteApi worker on close
When a worker is created for a flush of the worker on close, it will just
create a new object which makes it harder for for the garbage collector to
correctly free the unused objects because the Worker itself has a reference
to the WriteApi object it called. This is especially frustrating for
WriteType::SYNCHRONOUS which doesn't need workers at all.