fix: clear Client::autoCloseable on close
The autoCloseable list of a client is only growing but is never cleared.
This causes problems with the reference based garbage collection. But there
is nothing more to process when all writers were closed.

This problem can often be seen on a system that each call of
client->createWriteApi (on the same "global" client) followed by a
->write() will leak one or more file descriptor(s).